LayoutTests/js/get-by-pname-expected.txt

This tests that op_get_by_pname is compiled correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test tier: cold
PASS foo(o) is 11
PASS foo(p) is 10
PASS foo(q) is 3467
PASS foo(r) is 113
PASS foo(s) is 182

Test tier: llint
PASS foo(o) is 11
PASS foo(p) is 10
PASS foo(q) is 3467
PASS foo(r) is 113
PASS foo(s) is 182

Test tier: baseline
PASS foo(o) is 11
PASS foo(p) is 10
PASS foo(q) is 3467
PASS foo(r) is 113
PASS foo(s) is 182

Test tier: dfg
PASS foo(o) is 11
PASS foo(p) is 10
PASS foo(q) is 3467
PASS foo(r) is 113
PASS foo(s) is 182

PASS successfullyParsed is true

TEST COMPLETE
